Get HTTPS working in http_client by skipping over a super.
The super() used in http.client.HTTPSConnection leads to a bit of
recursion because of the already in place overrides.

To get around this we replace the __init__.

This unfortunately removes some SSL features, but on the other hand
it means that SSL type requests will work. For more effective testing
other libraries should be used.