Ensure PATH_INFO is escaped per PEP-333 implications
Fixes #27 which has a few more details on why PATH_INFO should be
unquoted before being set in the environment.

When doing so, though, we _must_ set either REQUEST_URI or RAW_URI in
the environ to make it possible to construct URIs that have had their
%2Fs trashed. From reading various post-wsgi documents REQUEST_URI is
most common, but some servers, such as gunicorn use RAW_URI, so since
its cheap, we use both here.

Thanks to @Vultaire for the bug report.