Fix handling of non-intercepted SSL
fixes #21

If an intercept is place but a domain that is non-intercepted
domain is requested, make a stronger effort to ensure that the
necessary pieces are in place for SSL to work correctly.

Tests adjusted or added for both httpclient and urllib, which
are the two libraries that had been failing in the past.