Intial stab at context manager based intercepts
See test/test_interceptor for the expect use pattern.

The idea here is that instead of the overrides being visible to the
caller the caller just uses a context manager that surrounds the tests
and that context manager does the right thing.

Those things, at the moment, are exploratory, in an effort to find the
right idiom. The use of classes is intentional, despite icky, to make
overrides and the creation of external subclasses easier.

One thing (of probably several) that remains to be determined is the best
idiom for interception when running a large suite of tests with the same
intercept? What controls the context entry and exit?

Nothing written in stone here, just an exploration. The goals? Two:

* make the external interface simple while allowing lots of changing
  underneath (to bring wsgi-intercept into the future)
* explore options for adding hooks and the like (see #28 for a
  related idea)