Guard against http_proxy and https_proxy
requests and urllib will get upset of socks proxy settings are
present in the environment. In those interceptors, an exception
will now be raised if $http_proxy or $https_proxy are set.

Tests are added to cover the variable being set for all four
of the interceptor types.

This is the quick and dirty solution to the problem. The way this
has been done shows some clear opportunities for refactoring down
the road.

Fixes #30