Add support for intercepting urllib3
This is relatively straightforward but presents a problem.
The monkeypatching is identical as that for requests, but on
different modules, because of the vendorization of urllib3
that requests does.

It would be better if there was some way to have just one
set of code that removes the duplication and does the right
but maintains the semantics of intercepting the desired thing
by name.