Don't use six when testing httplib
Otherwise monkeypatching get confused and will not work
and the tests will fail. An effort was made to use the
six http_client always (both in tests and the interceptor)
but this led to infinite recursion, so going for the lame
fix here.

The functionality has always worked (as long as six wasn't
involved) this is just getting things working in tests.