Ensure header values are native str
PEP 3333 requires that values in the environ that come from
header be native str objects decoded from ISO-8859-1 bytes.
This means that in Python 2 and 3 they are effectively
different things, but oh well, these are the rules.

One area where this causes a real issue is in the
SimpleCookie.load() method which checks for the type of the
string associated with HTTP_COOKIE.

Fixes #39