Switch to using tox for testing
This allows easier testing of multiple versions of Python in a more
automated fashion. This is needed if we're ever going to make some
of the planned extensive refactorings.

This change:

* adds tox.ini
* updates Makefile and .travis.yml to point to tox
* fixes a doctest typo found while make sure the Makefile targets
  work
* adds a [docs] extras_require so that the deps in tox are managed
  cleanly

The pep8 target is currently commented out because it fails (and
has done for a long time). Will fix asap. This commit is to ensure
that travis is working. Followups will tune it up.