Fix up README and version handling
With the addition of six to __init__.py the import of wsgi_intercept
in setup.py to get version and long description information no longer
works. Now we keep that information externally:

* version is static in setup.py
* the __doc__ from __init__.py is written to README.rst by a tox
  target, and that is loaded by setup.py

This has an added benefit of keeping the github visible readme,
the pypi info, and the main docs all in sync, at the cost of a
little magic. The __doc__ rarely changes, so the magic isn't
that bad.

Note that the bug in this code (the use of bytestrings) is not
yet fixed, but I'm making this change on that branch because it
makes the test usable.