Accept an empty string from a wsgi iterator
This is allowed by the pep but the code was written such that if it
received an empty string it would not get the rest of the iterator.
This change fixes it so only None will cause the stop.

With this change an unrelated test wsgi app which was returning an
empty string needed to be fixed to return a bytestring.

Fixes #46