Correct handling of a environ keys in python2
Both keys and values should be native str. This was
already the case for values, but not for keys. Thanks
to @nboullis for identifying the problem and the fix.

A test is added which runs a request against the wsgiref
sample app, this should help to protect against such
issues recurring.

Fixes #65